pyproject.toml
[tool.poetry]
name = "python-thingiverse"
version = "0.0.5rc1"
description = "A Python Thingiverse REST API wrapper"
authors = ["Jose Garcia <maton.pg@gmail.com>"]
packages = [
    { include = "thingiverse" }
]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
python-box = "^5.4.1"
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
tox = "^3.24.5"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
